Added User Roles and functionality to hide posts

+++ df2communitywebsite/app/Http/Controllers/PostController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Storage;
 use App\Post;
 
 
@@ -13,13 +14,19 @@ class PostController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
+    public function __construct()
+    {
+        $this->middleware('auth', ['except' => ['index', 'show']]);
+    }
+
     public function index()
     {
 //        $posts = Post::all();
-//        $posts = Post::orderBy('title', 'desc')->get();
+        $posts = Post::orderBy('title', 'desc')->get()->where('hidden', 'false');
 //        $posts = Post::orderBy('title', 'desc')->take(1)->get();
-//        $post = Post::where('title', 'Post Two')->get();
-        $posts = Post::orderBy('title', 'desc')->paginate(1);
+//        $posts = Post::where('title', 'Post Two')->get();
+//        $posts = Post::orderBy('title', 'desc')->get();
+//        $posts = Post::orderBy('title', 'desc')->paginate(1);
         return view('posts.index')->with('posts', $posts);
     }
 
@@ -44,12 +51,34 @@ class PostController extends Controller
         $this->validate($request, [
            'title' => 'required',
             'body' => 'required',
+            'cover_image' => 'image|nullable|max:1999',
+            'hide' => 'required',
         ]);
+
+        //Handle file upload
+        if($request->hasFile('cover_image'))
+        {
+            $fileNameWithExt = $request->file('cover_image')->getClientOriginalName();
+            //Get just filename
+            $fileName = pathinfo($fileNameWithExt, PATHINFO_FILENAME);
+            //Get extension
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+            //Filename to store
+            $fileNameToStore = $fileName. '_' .time(). '.' .$extension;
+            //Upload Image
+            $path = $request->file('cover_image')->storeAs('public/cover_images', $fileNameToStore);
+
+        }else{
+            $fileNameToStore = 'noimage.jpg';
+        }
 
         //create post
         $post = new Post;
         $post->title = $request->input('title');
         $post->body = $request->input('body');
+        $post->user_id = auth()->user()->id;
+        $post->cover_image = $fileNameToStore;
+        $post->hidden = $request->input('hide');
         $post->save();
 
         return redirect('/posts')->with('success', 'Post Created');
@@ -76,6 +105,12 @@ class PostController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
+
+        //check for correct user
+        if(auth()->user()->id !== $post->user_id){
+            return redirect('/posts')->with('error', "Unauthorized Page");
+        }
+
         return view('posts.edit')->with('post', $post);
     }
 
@@ -91,12 +126,41 @@ class PostController extends Controller
         $this->validate($request, [
             'title' => 'required',
             'body' => 'required',
+            'cover_image' => 'image|nullable|max:1999',
+            'hide' => 'required',
         ]);
 
-        //create post
+
         $post = Post::find($id);
+        //Handle file upload
+        if($request->hasFile('cover_image'))
+        {
+            $fileNameWithExt = $request->file('cover_image')->getClientOriginalName();
+            //Get just filename
+            $fileName = pathinfo($fileNameWithExt, PATHINFO_FILENAME);
+            //Get extension
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+            //Filename to store
+            $fileNameToStore = $fileName. '_' .time(). '.' .$extension;
+            //Upload Image
+            $path = $request->file('cover_image')->storeAs('public/cover_images', $fileNameToStore);
+
+            if($post->cover_image != 'noimage.jpg') {
+                Storage::delete('public/cover_images/' . $post->cover_image);
+            }
+            $post->cover_image = $fileNameToStore;
+
+        }
+
+
+        //create post
         $post->title = $request->input('title');
         $post->body = $request->input('body');
+        $post->hidden = $request->input('hide');
+        if($request->hasFile('cover_image')){
+            $post->cover_image = $fileNameToStore;
+        }
+
         $post->save();
 
         return redirect('/posts')->with('success', 'Post Updated');
@@ -111,8 +175,18 @@ class PostController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
-        $post->delete();
 
+        //check for correct user
+        if(auth()->user()->id !== $post->user_id){
+            return redirect('/posts')->with('error', "Unauthorized Page");
+        }
+
+        if($post->cover_image != 'noimage.jpg'){
+            //Delete the image
+            Storage::delete('public/cover_images/'.$post->cover_image);
+        }
+
+        $post->delete();
         return redirect('/posts')->with('success', 'Post Removed');
     }
 }
